Broken implementation with pointers
Does not compile any more, as swapCharacters has issues related to type-checking that prevent it from compiling, but should handle the data in a much safer way than before, preventing any out-of-bounds errors while still pushing any changes to the string on to the original string. Decided to push changes to original string instead of making a copy of the string to manipulate as we don't have any further need to use the original string anyways.